Attached and Detached Bug Fix
There was an issue with local vs global space and how objects were remaining attached to the
Gameworld and not their parent objects.

The initial goal for this commit was to fix the direction of particles on the player but that will have to be dealt with later.

I need to make a video.